requirements.txt
python-gnupg>=0.3.7
beautifulsoup4>=4.4
six